+++ src/model/starboardMessage.ts
